CPR-1141 do database work inside a transaction
